Clean up

+++ tpu_inference/kernels/megablox/gmm.py
@@ -510,9 +510,7 @@ def gmm(
                     # NOTE(catswe): chain-casting as direct uint4
                     # to bf16 for awq isn't yet supported in pallas
                     loaded_rhs_block = loaded_rhs_block.astype(
-                        jnp.int8).astype(
-                            jnp.bfloat16) - rhs_zero_point[b_i].astype(
-                                jnp.int8).astype(jnp.bfloat16)
+                        jnp.int8).astype(jnp.bfloat16) - rhs_zero_point[b_i]
 
                 partial_result = jnp.dot(
                     loaded_lhs_block,
